IOS-7613: Fix navbar appearance for iOS 15
